Added a base VPC setup. This will provision a total of 4 subnets (2 private, 2 public) in 2 availability zones, with all required options for running EC2 instances in them.
This will also start two NAT instances (one in each AZ) that will allow EC2 instances that will be added to the private subnets to communicate with the internet. For the purposes of this demonstration the NAT instances will also serve as jumpnodes to SSH into the private EC2 instances, which should be avoided in production setups.